+++ src/imio/smartweb/core/contents/sections/events/content.py
@@ -46,19 +46,19 @@ class ISectionEvents(ISection):
         required=True,
     )
 
-    max_nb_results = schema.Int(
-        title=_(u"Maximum number of items to display"),
-        required=True,
-        default=6,
-        min=1,
-        max=12,
-    )
-
     nb_results_by_batch = schema.Choice(
         title=_(u"Number of items per batch"),
         required=True,
         default=3,
         values=[1, 3, 4],
+    )
+
+    max_nb_batches = schema.Int(
+        title=_(u"Maximum number of batches to display"),
+        required=True,
+        default=2,
+        min=1,
+        max=12,
     )
 
     link_text = schema.TextLine(